Avoid declaration of more size_t variables

+++ 0x1D-search_algorithms/102-interpolation.c
@@ -13,7 +13,7 @@
  */
 int interpolation_search(int *array, size_t size, int value)
 {
-	size_t pos, lo, hi;
+	int pos, lo, hi;
 
 	if (array == NULL || size < 1)
 		return (-1);
@@ -25,24 +25,24 @@ int interpolation_search(int *array, size_t size, int value)
 	{
 		pos = lo + (((double)(hi - lo) / (array[hi] - array[lo])) *
 				(value - array[lo]));
-		if (pos >= size)
+		if (pos >= (int) size)
 		{
-			printf("Value checked array[%zd] is out of range\n", pos);
+			printf("Value checked array[%d] is out of range\n", pos);
 			return (-1);
 		}
-		printf("Value checked array[%zd] = [%d]\n", pos, array[pos]);
+		printf("Value checked array[%d] = [%d]\n", pos, array[pos]);
 
 		if (value == array[pos])
-			return ((int) pos);
+			return (pos);
 		else if (value > array[pos])
 			lo = pos + 1;
 		else
 			hi = pos - 1;
 	}
 
-	printf("Value checked array[%zd] = [%d]\n", lo, array[lo]);
+	printf("Value checked array[%d] = [%d]\n", lo, array[lo]);
 	if (value == array[lo])
-		return ((int) lo);
+		return (lo);
 	else
 		return (-1);
 }
